+++ src/data/teamMembers.js
@@ -34,7 +34,7 @@ export const teamMembers = [
         surname: "Dr. Ocampo",
         photo: "https://via.placeholder.com/150",
         title: "Médico Veterinario",
-        specialty: "Clínica e Histopatología",
+        specialty: "Clínica y Citología",
         license: "M.P N° 1221",
         description:
             "Se especializa en anatomo-histopatología y citología, además de clínica general, manejando análisis citológicos de mascotas.",
@@ -54,7 +54,7 @@ export const teamMembers = [
         surname: "Dra. Vallejos",
         photo: "https://via.placeholder.com/150",
         title: "Médica Veterinaria",
-        specialty: "Clínica y Nutrición",
+        specialty: "Nutrición",
         license: "M.P N° 1298",
         description:
             "Se especializa en nutrición de mascotas, colaborando en dietas y planes alimenticios personalizados para perros y gatos.",
